+++ pages/pug.vue
@@ -1,7 +1,7 @@
 <template lang="pug">
   div.pug
     section.section
-      .container
+      .container.mx-auto
         article#html.content.is-medium(v-html="pug")
 </template>
 
@@ -18,6 +18,11 @@ export default {
       title: [this.title, this.subtitle].join(' - '),
     }
   },
+  head() {
+    return {
+      title: this.title,
+    }
+  },
   computed: {
     pug() {
       return pug
@@ -26,15 +31,10 @@ export default {
   async mounted() {
     this.title = (await this.pageTitle) + ' - ' + this.subtitle
   },
-  head() {
-    return {
-      title: this.title,
-    }
-  },
 }
 </script>
 
-<style lang="scss">
+<style>
 .section > .container {
   background-color: white;
   padding: 0 1em;